+++ app/[blog_name]/(category)/[category_id]/page.jsx
