+++ app/[blog_name]/layout.jsx
@@ -1,0 +1,12 @@
+import "../page.module.css";
+
+import React from "react";
+
+export default function blogLayout({ children }) {
+  return (
+    <>
+      <div>{children}</div>
+      <div>lay</div>
+    </>
+  );
+}
