+++ app/[blog_name]/manage/layout.jsx
@@ -1,0 +1,10 @@
+import "../../page.module.css";
+
+export default function manageLayout({ children }) {
+  return (
+    <>
+      <div>{children}</div>
+      <div>lay2</div>
+    </>
+  );
+}
