+++ app/[blog_name]/manage/posts/page.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function posts() {
+  return <div>page</div>;
+}
