+++ app/auth/login/page.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import LoginView from "../../../components/LoginView";
+
+export default function login() {
+  return (
+    <>
+      <LoginView />
+    </>
+  );
+}
