+++ app/loading.jsx
@@ -1,0 +1,5 @@
+import LoadingSkeleton from "../components/LoadingSkeleton";
+export default function Loading() {
+  // You can add any UI inside Loading, including a Skeleton.
+  return <LoadingSkeleton />;
+}
