+++ app/user/account/page.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function page() {
+  return <div>page</div>;
+}
