+++ app/user/myBlogs/page.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import BlogCreate from "../../../../components/BlogCreate";
+
+export default function myBlogs() {
+  return (
+    <>
+      <BlogCreate />
+    </>
+  );
+}
