+++ components/LoadingSkeleton.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function LoadingSkeleton() {
+  return <div>LoadingSkeleton</div>;
+}
