--- src/app/[blog_id]/(category)/[category_id]/[article_id]/page.js
@@ -1,3 +1,0 @@
-export default function blogs({}) {
-  return <div>blogs</div>;
-}
