--- src/app/auth/login/page.js
@@ -1,11 +1,0 @@
-import LoginView from "@/components/LoginView";
-
-import React from "react";
-
-export default function login() {
-  return (
-    <>
-      <LoginView />
-    </>
-  );
-}
